requirements.txt
flask
flask-login
flask-migrate
flask-sqlalchemy
mysql-connector-python
python-dotenv